Make standalone sandbox egress fail closed
