+++ .env.example
@@ -157,13 +157,8 @@ OVERLAY_VERCEL_SANDBOX_CREATION_USD=0.0000006
 OVERLAY_VERCEL_SANDBOX_RESERVATION_BUFFER_PERCENT=25
 OVERLAY_SANDBOX_MAX_PROVIDER_COST_USD_PER_RUN=15
 OVERLAY_SANDBOX_PROVIDER_SPEND_ALERT_USD=10
-# Standalone sandbox networking is denied by default. Enabling it requires an
-# explicit comma-separated Vercel domain allowlist; private/link-local CIDRs
-# remain denied. Example: registry.npmjs.org,pypi.org,files.pythonhosted.org
-OVERLAY_SANDBOX_RUN_ALLOWED_DOMAINS=
-# The command is terminated when aggregate non-loopback transmit bytes exceed
-# this limit. Reservation also includes the separately bounded artifact reads.
-OVERLAY_SANDBOX_RUN_MAX_EGRESS_BYTES=10000000000
+# Standalone API sandbox networking is always denied. Networked agent workflows
+# use separate, explicitly metered sandbox paths.
 VERCEL_TOKEN
 VERCEL_TEAM_ID
 VERCEL_PROJECT_ID
